Multiple changes to SPI Variants
Created custom elements for GDS components not in in this proto. ver.
Elements above were the Warning and Important block elements.
Reverted Important block to Warning blocks for the purpose of testing.
Tidied variations of Warning to be one single block with Nunjucks.
Corrected supplementary text for Adjusted Discretionary Bursary Fund.